encapsulate resolvers and service factories
