+++ packages/tools/webpack-component-loader/src/loader.ts
@@ -12,23 +12,15 @@ import {
     isFluidPackage,
 } from "@microsoft/fluid-container-definitions";
 import { Container } from "@microsoft/fluid-container-loader";
-import { IDocumentServiceFactory, IUrlResolver } from "@microsoft/fluid-driver-definitions";
-import { TestDocumentServiceFactory, TestResolver } from "@microsoft/fluid-local-driver";
-import { ILocalDeltaConnectionServer, LocalDeltaConnectionServer } from "@microsoft/fluid-server-local-server";
-import { SessionStorageDbFactory } from "@microsoft/fluid-local-test-utils";
+import { IDocumentServiceFactory } from "@microsoft/fluid-driver-definitions";
 import { IUser } from "@microsoft/fluid-protocol-definitions";
-import { DefaultErrorTracking, RouterliciousDocumentServiceFactory } from "@microsoft/fluid-routerlicious-driver";
 import { getRandomName } from "@microsoft/fluid-server-services-client";
 import { extractDetails, IResolvedPackage } from "@microsoft/fluid-web-code-loader";
 import { Deferred } from "@microsoft/fluid-common-utils";
-import * as jwt from "jsonwebtoken";
-// eslint-disable-next-line import/no-internal-modules
-import * as uuid from "uuid/v4";
-import { OdspDocumentServiceFactory, IOdspNewFileParams } from "@microsoft/fluid-odsp-driver";
 import { HTMLViewAdapter } from "@microsoft/fluid-view-adapters";
-import { InsecureUrlResolver } from "@microsoft/fluid-test-runtime-utils";
-import { IComponent, IRequest } from "@microsoft/fluid-component-core-interfaces";
-import { OdspUrlResolver } from "./odspUrlResolver";
+import { IComponent } from "@microsoft/fluid-component-core-interfaces";
+import { MultiUrlResolver } from "./multiResolver";
+import { getDocumentServiceFactory } from "./multiDocumentServiceFactory";
 
 export interface IDevServerUser extends IUser {
     name: string;
@@ -81,10 +73,6 @@ export type RouteOptions =
     | ITinyliciousRouteOptions
     | IOdspRouteOptions;
 
-const getUser = (): IDevServerUser => ({
-    id: uuid(),
-    name: getRandomName(),
-});
 
 async function loadScripts(files: string[], origin: string) {
     // Add script to page, rather than load bundle directly
@@ -174,85 +162,6 @@ async function getResolvedPackage(
     };
 }
 
-function getUrlResolver(
-    documentId: string,
-    options: RouteOptions,
-): IUrlResolver {
-    switch (options.mode) {
-        case "docker":
-            return new InsecureUrlResolver(
-                "http://localhost:3000",
-                "http://localhost:3003",
-                "http://localhost:3001",
-                options.tenantId,
-                options.tenantSecret,
-                getUser(),
-                options.bearerSecret);
-
-        case "r11s":
-            return new InsecureUrlResolver(
-                options.fluidHost,
-                options.fluidHost.replace("www", "alfred"),
-                options.fluidHost.replace("www", "historian"),
-                options.tenantId,
-                options.tenantSecret,
-                getUser(),
-                options.bearerSecret);
-
-        case "tinylicious":
-            return new InsecureUrlResolver(
-                "http://localhost:3000",
-                "http://localhost:3000",
-                "http://localhost:3000",
-                "tinylicious",
-                "12345",
-                getUser(),
-                options.bearerSecret);
-
-        case "spo":
-        case "spo-df":
-            return new OdspUrlResolver(
-                options.server,
-                { accessToken: options.odspAccessToken });
-
-        default: // Local
-            return new TestResolver(documentId);
-    }
-}
-
-function createRequestForCreateNew(
-    rawUrl: string,
-    options: RouteOptions,
-    resolver: IUrlResolver,
-    fileName: string,
-): IRequest {
-    let request: IRequest;
-    switch (options.mode) {
-        case "r11s":
-            request = (resolver as InsecureUrlResolver).createCreateNewRequest(rawUrl, fileName);
-            break;
-        case "spo":
-        case "spo-df":
-            const params: IOdspNewFileParams = {
-                fileName,
-                driveId: options.driveId,
-                filePath: "/r11s/",
-                siteUrl: `https://${options.server}`,
-            };
-            request = (resolver as OdspUrlResolver).createCreateNewRequest(rawUrl, params);
-            break;
-        case "docker":
-            request = (resolver as InsecureUrlResolver).createCreateNewRequest(rawUrl, fileName);
-            break;
-        case "tinylicious":
-            request = (resolver as InsecureUrlResolver).createCreateNewRequest(rawUrl, fileName);
-            break;
-        default: // Local
-            request = (resolver as TestResolver).createCreateNewRequest();
-    }
-    return request;
-}
-
 // Invoked by `start()` when the 'double' option is enabled to create the side-by-side panes.
 function makeSideBySideDiv(divId?: string) {
     const div = document.createElement("div");
@@ -265,43 +174,6 @@ function makeSideBySideDiv(divId?: string) {
         div.id = divId;
     }
     return div;
-}
-
-function getDocumentServiceFactory(documentId: string, options: RouteOptions) {
-    let documentServiceFactory: IDocumentServiceFactory;
-    let deltaConn: ILocalDeltaConnectionServer;
-    switch (options.mode) {
-        case "local": {
-            deltaConn = LocalDeltaConnectionServer.create(new SessionStorageDbFactory(documentId));
-            documentServiceFactory = new TestDocumentServiceFactory(deltaConn);
-            break;
-        }
-        case "spo":
-        case "spo-df": {
-            // TODO: web socket token
-            documentServiceFactory = new OdspDocumentServiceFactory(
-                "webpack-component-loader",
-                async (siteUrl, refresh) => { return options.odspAccessToken; },
-                async (refresh) => { return options.pushAccessToken; },
-                { send: (event) => { return; } },
-                undefined,
-                undefined,
-                undefined,
-                options.openMode === "detached" ? true : false,
-            );
-            break;
-        }
-        default: {
-            documentServiceFactory = new RouterliciousDocumentServiceFactory(
-                false,
-                new DefaultErrorTracking(),
-                false,
-                true,
-                undefined,
-            );
-        }
-    }
-    return {documentServiceFactory, connection: deltaConn};
 }
 
 export async function start(
@@ -324,12 +196,11 @@ export async function start(
         textArea.style.display = "none";
     }
 
-    const {documentServiceFactory, connection} = getDocumentServiceFactory(finalDocId, options);
-
-    const urlResolver = getUrlResolver(finalDocId, options);
+    const documentServiceFactory = getDocumentServiceFactory(finalDocId, options);
 
     // Construct a request
     const url = window.location.href;
+    const urlResolver = new MultiUrlResolver(url, documentId, options);
 
     // Create Package
     const scriptIds: string[] = [];
@@ -367,7 +238,7 @@ export async function start(
         // eslint-disable-next-line @typescript-eslint/no-floating-promises
         urlDeferred.promise.then(async (containerUrl) => {
             // New documentServiceFactory for right div, same everything else
-            const docServFac2: IDocumentServiceFactory = new TestDocumentServiceFactory(connection);
+            const docServFac2: IDocumentServiceFactory = getDocumentServiceFactory(finalDocId, options);
             const hostConf2 = { documentServiceFactory: docServFac2, urlResolver };
 
             // This will create a new Loader/Container/Component from the BaseHost above. This is
@@ -389,7 +260,7 @@ export async function start(
     }
     if (!attached) {
         attachButton.onclick = async () => {
-            await container1.attach(createRequestForCreateNew(window.location.href, options, urlResolver, finalDocId))
+            await container1.attach(urlResolver.createRequestForCreateNew(finalDocId))
                 .then(() => {
                     const text = window.location.href.replace("create", finalDocId);
                     textArea.innerText = text;
@@ -415,7 +286,7 @@ async function startRendering(container: Container, url: string, div: HTMLDivEle
             }).catch((error) => reject(error));
         };
 
-        container.on("contextChanged", (value) => {
+        container.on("contextChanged", () => {
             tryGetComponentAndRender();
         });
         tryGetComponentAndRender();
@@ -443,10 +314,4 @@ async function getComponentAndRender(container: Container, url: string, div: HTM
     const view = new HTMLViewAdapter(component);
     view.render(div, { display: "block" });
     return true;
-}
-
-export function getUserToken(bearerSecret: string) {
-    const user = getUser();
-
-    return jwt.sign({ user }, bearerSecret);
 }
